Add signOrder and getOrderHashHex
getOrderHashHex generates the hex encoded hash for an order, and
signOrder generates the ecSignature for that order and returns a
promise for the signed order.

Both getOrderHashHex() and signOrder() can take either an order, or
a promise for an order, and return a promise.

From a testing perspective, verifying a hash requires setting a
pre-determined salt. When testing signatures, we use a pre-determined
salt to test calling signOrder() on an order object, and use ZeroEx
to verify the hash signature when calling signOrder() on a promise.