Add setMakerAllowances
Given an order, setMakerAllowances will make sure allowances are
set on the account such that the order can be completed.

I had to downgrade testrpc due to an issue with gas estimates in
the 0x.js library. If / when https://github.com/0xProject/0x.js/pull/199
is merged we can get back up-to-date.

This also adjusts the tests to reset after each test, instead of
at the end of all tests.