Update 0x.js to v 0.27.2
There were serveral changes between 0.22.6 and 0.27.2 which had to
be addressed in the codebase to update. In general I tried to take
the least invasive approach to updating things, which sometimes lead
to promisifying things that 0x.js just unpromisified.

Also, Now openrelay.js has a "ready" promise that needs to be resolved
before the zeroEx attribute is usable. The openrelay methods will resolve
this automatically, but accessing the zeroEx attribute directly puts that
onus on the client code.