Update testrpc to 6.0.3
The new version of 0x.js uses an updated testrpc, with a corresponding
chaindb dump. This meant that some contracts had moved around (such
as the ZRX token contract), and using the old db dump with the new
0x.js didn't work. This update uses a new testrpc version and updates
the testrpc dump.

Also, for some reason on this version of testrpc a bunch of transactions
were failing, saying that the transaction gas exceeded the block limit.
TestRPC has a pretty low default gas limit, so I've bumped it up since
many of the transactions exceeded the default.